:hammer: chore: refactor some API folders
