+++ packages/api/src/controllers/admin/manage/cad-settings/discord/DiscordSettingsController.ts
@@ -8,7 +8,7 @@ import type { cad } from "@prisma/client";
 import { BadRequest } from "@tsed/exceptions";
 import { DISCORD_SETTINGS_SCHEMA } from "@snailycad/schemas";
 import { validateSchema } from "lib/validateSchema";
-import { getRest } from "lib/discord";
+import { getRest } from "lib/discord/config";
 
 const guildId = process.env.DISCORD_SERVER_ID;
