+++ packages/api/src/lib/discord/config.ts
@@ -1,8 +1,8 @@
 import process from "process";
 import { REST } from "@discordjs/rest";
 
-export const DISCORD_API_VERSION = "v10";
-export const DISCORD_API_URL = `https://discord.com/api/${DISCORD_API_VERSION}`;
+export const DISCORD_API_VERSION = "10" as const;
+export const DISCORD_API_URL = `https://discord.com/api/v${DISCORD_API_VERSION}`;
 export const GUILD_ID = process.env.DISCORD_SERVER_ID;
 export const BOT_TOKEN = process.env.DISCORD_BOT_TOKEN;
 
@@ -12,7 +12,7 @@ export function getRest(): REST {
     throw new Error("mustSetBotTokenGuildId");
   }
 
-  cacheREST ??= new REST({ version: "10" }).setToken(BOT_TOKEN);
+  cacheREST ??= new REST({ version: DISCORD_API_VERSION }).setToken(BOT_TOKEN);
 
   if (process.env.NODE_ENV === "development") {
     cacheREST.on("restDebug", console.log);
